% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sharedobe_function.R
\name{sharedobe_func}
\alias{sharedobe_func}
\title{My doggie Function}
\usage{
sharedobe_func(love = TRUE)
}
\arguments{
\item{love}{Do you love your dog? Defaults to TRUE.}
}
\description{
This function allows you to express my dog's love .
}
\examples{
sharedobe_func()
}
\keyword{dog}

